+++ src/generator.rs
@@ -225,33 +225,20 @@ impl<R: rand::RngCore> Scru128Generator<R> {
     ///
     /// Note that the generator object should be protected from concurrent accesses during the
     /// sequential calls to a generation method and this method to avoid race conditions.
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// # #[cfg(feature = "std")]
-    /// # {
-    /// use scru128::generator::{Scru128Generator, Status};
-    ///
-    /// let mut g = Scru128Generator::new();
-    /// let x = g.generate();
-    /// let y = g.generate();
-    /// if g.last_status() == Status::ClockRollback {
-    ///     panic!("clock moved backward");
-    /// } else {
-    ///     assert!(x < y);
-    /// }
-    /// # }
-    /// ```
+    #[deprecated(
+        since = "2.6.0",
+        note = "use `generate_or_abort()` to guarantee monotonicity"
+    )]
     pub const fn last_status(&self) -> Status {
         self.last_status
     }
 }
 
-/// The status code returned by [`Scru128Generator::last_status()`] method.
-#[derive(Copy, Clone, Eq, PartialEq, Hash, Debug)]
+/// _Deprecated_: The status code returned by [`Scru128Generator::last_status()`] method.
+#[derive(Copy, Clone, Eq, PartialEq, Hash, Debug, Default)]
 pub enum Status {
     /// Indicates that the generator has yet to generate an ID.
+    #[default]
     NotExecuted,
 
     /// Indicates that the latest `timestamp` was used because it was greater than the previous
@@ -272,12 +259,6 @@ pub enum Status {
     /// Indicates that the monotonic order of generated IDs was broken because the latest
     /// `timestamp` was less than the previous one by ten seconds or more.
     ClockRollback,
-}
-
-impl Default for Status {
-    fn default() -> Self {
-        Status::NotExecuted
-    }
 }
 
 #[cfg(feature = "std")]
@@ -357,18 +338,18 @@ mod tests_generate_or_reset {
     fn generates_increasing_ids_even_with_decreasing_or_constant_timestamp() {
         let ts = 0x0123_4567_89abu64;
         let mut g = Scru128Generator::new();
-        assert_eq!(g.last_status(), Status::NotExecuted);
+        assert_eq!(g.last_status, Status::NotExecuted);
 
         let mut prev = g.generate_or_reset_core(ts, 10_000);
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
         assert_eq!(prev.timestamp(), ts);
 
         for i in 0..100_000u64 {
             let curr = g.generate_or_reset_core(ts - i.min(9_998), 10_000);
             assert!(
-                g.last_status() == Status::CounterLoInc
-                    || g.last_status() == Status::CounterHiInc
-                    || g.last_status() == Status::TimestampInc
+                g.last_status == Status::CounterLoInc
+                    || g.last_status == Status::CounterHiInc
+                    || g.last_status == Status::TimestampInc
             );
             assert!(prev < curr);
             prev = curr;
@@ -381,23 +362,23 @@ mod tests_generate_or_reset {
     fn breaks_increasing_order_of_ids_if_timestamp_moves_backward_a_lot() {
         let ts = 0x0123_4567_89abu64;
         let mut g = Scru128Generator::new();
-        assert_eq!(g.last_status(), Status::NotExecuted);
+        assert_eq!(g.last_status, Status::NotExecuted);
 
         let mut prev = g.generate_or_reset_core(ts, 10_000);
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
         assert_eq!(prev.timestamp(), ts);
 
         let mut curr = g.generate_or_reset_core(ts - 10_000, 10_000);
-        assert_eq!(g.last_status(), Status::ClockRollback);
+        assert_eq!(g.last_status, Status::ClockRollback);
         assert!(prev > curr);
         assert_eq!(curr.timestamp(), ts - 10_000);
 
         prev = curr;
         curr = g.generate_or_reset_core(ts - 10_001, 10_000);
         assert!(
-            g.last_status() == Status::CounterLoInc
-                || g.last_status() == Status::CounterHiInc
-                || g.last_status() == Status::TimestampInc
+            g.last_status == Status::CounterLoInc
+                || g.last_status == Status::CounterHiInc
+                || g.last_status == Status::TimestampInc
         );
         assert!(prev < curr);
     }
@@ -413,18 +394,18 @@ mod tests_generate_or_abort {
     fn generates_increasing_ids_even_with_decreasing_or_constant_timestamp() {
         let ts = 0x0123_4567_89abu64;
         let mut g = Scru128Generator::new();
-        assert_eq!(g.last_status(), Status::NotExecuted);
+        assert_eq!(g.last_status, Status::NotExecuted);
 
         let mut prev = g.generate_or_abort_core(ts, 10_000).unwrap();
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
         assert_eq!(prev.timestamp(), ts);
 
         for i in 0..100_000u64 {
             let curr = g.generate_or_abort_core(ts - i.min(9_998), 10_000).unwrap();
             assert!(
-                g.last_status() == Status::CounterLoInc
-                    || g.last_status() == Status::CounterHiInc
-                    || g.last_status() == Status::TimestampInc
+                g.last_status == Status::CounterLoInc
+                    || g.last_status == Status::CounterHiInc
+                    || g.last_status == Status::TimestampInc
             );
             assert!(prev < curr);
             prev = curr;
@@ -437,19 +418,19 @@ mod tests_generate_or_abort {
     fn returns_none_if_timestamp_moves_backward_a_lot() {
         let ts = 0x0123_4567_89abu64;
         let mut g = Scru128Generator::new();
-        assert_eq!(g.last_status(), Status::NotExecuted);
+        assert_eq!(g.last_status, Status::NotExecuted);
 
         let prev = g.generate_or_abort_core(ts, 10_000).unwrap();
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
         assert_eq!(prev.timestamp(), ts);
 
         let mut curr = g.generate_or_abort_core(ts - 10_000, 10_000);
         assert!(curr.is_none());
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
 
         curr = g.generate_or_abort_core(ts - 10_001, 10_000);
         assert!(curr.is_none());
-        assert_eq!(g.last_status(), Status::NewTimestamp);
+        assert_eq!(g.last_status, Status::NewTimestamp);
     }
 }